Adds an endpoint for API based authentication
The endpoints are POST /api/auth and DELETE /api/auth

For the signin (POST /api/auth) send a JSON payload of:

{
  "user": {
    "email": "...."
    "password": "..."
  }
}

You will either get a 200 OK response with the following body:

{ "ok": true }

or you will get a 401 Unauthorized response with no body.

If successful, the JWT token is in the response headers in the
`Authorization` header field.